fix: typos and rename models and controllers
