+++ .gitignore
@@ -6,3 +6,4 @@
 
 # env
 .env
+*.log
